SOMC

Description

This package exports functions to group and manipulate points in multi-dimensional space using self-organizing memory cards.

To understand the mechanics of most functions, it is necessary to understand how a self-organizing memory card works, at least how it works in this project.

Understanding Self-Organizing Memory Cards

What you want to do, and what you need for it

You want to group data. For this reason, you obviously need two components: Data, and you have to know how many groups you want.

But before you can group your data, you need a list of prototypes. Prototypes are essential to grouping data with memory cards, because they are the representation of groups. Thus you need as many prototypes as groups. Furthermore, prototypes contain a representation point in "data space" (space for data points) and a prototype position in "prototype space". It is a convention to initialize the data space with random values, while the prototypes are aligned tidily in prototype space.

Getting to the point

Because a self-organizing memory card is a type of neural network, it works by training. It is trained by going through every data point in your data set and moving the prototype closest to the data point a bit more in that direction. There is a problem with that: The resulting groups (prototypes) are not ordered at all. You can fix that by doing the following: If you move a prototype (lets call it "the winner"), you move all prototypes near the winner slightly to the direction of the point the winner moves to. And with "near" I mean close in another space, the prototype space! In the prototype space, all prototypes are aligned on a grid. Using this technique, prototypes next to each other in prototype space will always be next to each other in data space, too.

Basic Types

type Vec a = [a]

This type assignment is used to provide more intuitive working with vectors.

type DataPoint = Vec Double

A point in multi-dimensional space.

type DataSet = Set DataPoint

This type assignment is used for more intuitive typing, too.

data Prototype

A prototype in multi-dimensional space. It both uses a "represenation point" and a "position in prototype space". The "representation point" is (or should be) the middle of a group of data points, whereas the position in prototype space is used for regularly distributing representation points in data point space. This is done by moving prototypes close to each other in prototype space near to each other in data point space.

(-@>) :: [Double] -> [Int] -> Prototype

The prototype assignment operator just says "clip this data point to that prototype position".

type Prototypes = [Prototype]

This type assignment is used for more intuitive typing :).

type MDCS = (DataSet, Prototypes)

Multi-Dimensional Coordinate System. This type assignment is used to combine both data of the data space and prototypes of prototype space.

Functions for working with Vectors

(>+) :: Num a => Vec a -> Vec a -> Vec a

Add two vectors. This is defined as (>+) = zipWith (+).

(>-) :: Num a => Vec a -> Vec a -> Vec a

Subtract two vectors. This is defined as (>-) = zipWith (-).

(>*) :: Num a => Vec a -> a -> Vec a

Multiply vector by a scalar. This is defined as (>*) x y = map (y *) x.

absv :: Floating a => Vec a -> a

Length of vector, calculated using the euclidean distance.

Working with Multi-Dimensional Coordinate Systems

Using Self-Organizing Memory Cards

getWinner

Arguments

:: (DataPoint -> DataPoint -> Double)
A function which takes the position of the data point and the representation point of a prototype and returns the distance of both ("distance function").

-> Prototypes
The prototypes to choose from.

-> DataPoint
The data point to compare with.

-> Prototype
The "winning" prototype.

Returns the prototype for which the given function returns the lowest value.

updateWinner

Arguments

:: (Vec Int -> Vec Int -> Double)
The radius function. It takes the position of the winner prototype and the position of any prototype in prototype space and returns how much (relative) the prototype should be moved.

-> Prototypes
The prototype list to choose from.

-> Prototype
The winning prototype.

-> DataPoint
The data point to in which direction the prototypes should be moved.

-> Prototypes
The modified prototype list.

Updates a prototype given a "radius function" (a distance function for prototype space), a list of available prototypes, the "winner" prototype and a data point. Essentially, this function first generates a vector from any prototype to the given data point and then moves the prototype by a fraction of that vector, whereas the fraction is given by the radius function.

epoch

Arguments

:: (DataPoint -> DataPoint -> Double)
The distance function, as discussed with the getWinner function.

-> (Vec Int -> Vec Int -> Double)
The radius function, as discussed with the updateWinner function.

-> MDCS
The MDCS containing data points and prototypes.

-> Prototypes
The trained prototype list.

Trains the self-organizing memory card in one "epoch". An epoch is a training cycle, so for every given data point, the nearest prototype is determined (using the getWinner function) and then all prototypes are updated (based on the updateWinner function).

train

Arguments

:: (DataPoint -> DataPoint -> Double)
distance function

-> (Int -> Int -> Vec Int -> Vec Int -> Double)
Radius function, enhanced by two arguments: The first one is the amount of training cycles, the second is the current training cycle (as the output of this function should change over time).

-> MDCS

-> Int
amount of training cycles

-> Prototypes

A function for training a SOM. It iterates the epoch-function.

Processing Multi-Dimensional Coordinate Systems

sortToGroups

Arguments

:: (DataPoint -> DataPoint -> Double)
The distance function, as discussed with the getWinner function.

-> MDCS
The MDCS containing data points to sort and prototypes to sort to.

-> [(Prototype, DataSet)]
The data points of the data set of the MDCS, assigned to the protoypes.

This function sorts the data points of a MDCS to groups by going through every data point in the DataSet and then assigning it to the prototype which representation point is the closest to the data point.

Auxiliary Functions

The following functions can be used to build customized training functions. It is important to understand how they work together, so they can be built together correctly. train and epoch both use alpha-functions which tell how far a prototype has to be moved. As the names of quadraticAlpha, reversedAlpha and linearAlpha suggest, these are the most important parts of configuration. Their types are equal. They all take a radius function , a radius-based alpha function and a initial alpha value and return a configured alpha function. Radius functions are constructed using squareRadius or similar functions. Radius-based alpha functions are constructed using an interpretation, for example squareInterpretation. But an interpretation also needs to know how the alpha value should change with greater radius, so it needs a radius-based alpha distribution. So a complete alpha function could look like quadraticAlpha (squareRadius 3) (hexagonalInterpretation squareRadiusDistribution) 0.3.

printSet :: Show a => Set a -> IO ()

Prints a set in a neat way by just printing every element, so the developer can see what the set contains in an easy way.

printList :: Show a => [a] -> IO ()

Works the same es printSet, just with lists.

euclidDistance :: DataPoint -> DataPoint -> Double

Calculates the euclidean distance between two points, using the absv function.

Base Alpha Functions

quadraticAlpha

Arguments

:: (Int -> Int -> Int)

-> (Int -> Double -> Vec Int -> Vec Int -> Double)
A function which takes a radius and an alpha value and returns a radius-based alpha function.

-> Double
A initial alpha value. It decreases quadratically, which means it starts to decrease slowly and then faster.

-> Int

-> Int

-> Vec Int

-> Vec Int

-> Double

A epoch-based alpha distribution using quadratic functions for underlying calculations.

reversedAlpha :: (Int -> Int -> Int) -> (Int -> Double -> Vec Int -> Vec Int -> Double) -> Double -> Int -> Int -> Vec Int -> Vec Int -> Double

This epoch-based alpha distribution is unique, because it does not depend on the amount of training epochs, but rather decreases like 1/x. The more training samples there are, the more accurate are the results.

linearAlpha :: (Int -> Int -> Int) -> (Int -> Double -> Vec Int -> Vec Int -> Double) -> Double -> Int -> Int -> Vec Int -> Vec Int -> Double

This epoch-based alpha distribution uses a linearly decreasing alpha value.

Prototype Space Interpretations

squareInterpretation :: (Double -> Int -> Int -> Double) -> Int -> Double -> Vec Int -> Vec Int -> Double

A radius interpretation using squares as prototype "tiles".

independentInterpretation :: Double -> Vec Int -> Vec Int -> Double

A function taking an alpha value returning a radius-based alpha function, as discussed with the updateWinner function. independentRadius a w t returns a if t == w, and 0 if t /= w. In other words, only the winning prototype is updated with an alpha factor of a.

hexagonalInterpretation :: (Double -> Int -> Int -> Double) -> Int -> Double -> Vec Int -> Vec Int -> Double

A radius interpretation using hexagonal "tiles" for the prototypes.

Epoch-Based Radius Functions

squareRadius

Arguments

:: Int
starting radius

-> Int

-> Int

-> Int

A radius-based alpha distribution using a quadratic function for underlying calculations.

Radius-Based Alpha Functions

squareRadiusDistribution

Arguments

:: Double
A regulating variable between (!) 0 and (about) 0.5, stating how curvy the parabola should be. If this value is r and the following value is e, then the parabola goes through the points (0|1), (0.5|r) and (1|e), whereas 1 at the y-axis is the maximum radius possible and 1 at the x-axis is the maximum alpha possible.

-> Double
A regulating variable between 0 and the previous variable, stating how low the parabola should go at the maximum radius possible: If this value is called e and the maximum alpha value possible is a, then the outermost prototype is moved by e*a.

-> Double

-> Int

-> Int

-> Double
The final radius-based alpha distribution.

A distribution of the alpha value throughout different radii. It can be viewed as a graph of a parabola, whereas radii are noted on the x-axis and alpha values are noted on the y-axis.